chore: split sequences package
Splitting the modules along certain domains makes using the packages
easier, since it allows you to selectively use functionality.

This addresses the need for sequence types if you have a need of the
domain for types such as DNA, RNA, and Amino Acids. It also provides a
framework for defining almost any conversion between sequence types.

The side effect of this decision is a larger overhead in terms of
managing the project as a whole.