update: use charlist as sequence impl
Using the charlist as the underlying implementation really helps to
clean things up. I also was able to get rid of several non-essential
modules because of this.

There are a few places where I'd like to keep cleaning up the
implementations, but I'm happy with how this looks for now. I need to
review how I'm going to increment the versioning.