chore: update todos and random dna
Generating random dna is something that is very useful for e.g. fuzz testing. So
I wanted to have the generator play nicer with shell tooling. By outputing to
the stdio handle, we allow the output to be piped to wherever. This is much
nicer for writing scripts. Additionally, the simple newline approach allows me
to combine it with other tools for generating e.g. FASTA file headers.